parser/requirements.txt
PyYAML==6.0
Django==4.1.4
